move zip method to utils
