+++ src/main/utils/utils.ts
@@ -1,8 +1,14 @@
 import * as pathUtils from 'path';
 import * as vscode from 'vscode';
 import * as os from 'os';
+import * as fs from 'fs';
+import AdmZip, { IZipEntry } from 'adm-zip';
 
 export class Utils {
+    private static readonly MAX_FILES_EXTRACTED_ZIP: number = 1000;
+    private static readonly MAX_SIZE_EXTRACTED_ZIP: number = 1000000000; // 1 GB
+    private static readonly COMPRESSION_THRESHOLD_RATIO: number = 100;
+
     /**
      *  @returns the last segment of a path.
      * @param path
@@ -49,7 +55,73 @@ export class Utils {
         return str + '.zip';
     }
 
+    public static saveAsZip(zipPath: string, ...files: { fileName: string; content: string }[]): void {
+        let zip: AdmZip = new AdmZip();
+        for (let file of files) {
+            zip.addFile(file.fileName, Buffer.alloc(file.content.length, file.content));
+        }
+        zip.writeZip(zipPath);
+    }
+
+    static extractZip(zipPath: string, targetDir: string, overwrite: boolean = true, keepOriginalPermission: boolean = true): any {
+        if (!fs.existsSync(zipPath)) {
+            return '';
+        }
+        if (!fs.existsSync(targetDir)) {
+            fs.mkdirSync(targetDir, { recursive: true });
+        }
+        let zip: AdmZip = new AdmZip(zipPath);
+        let fileCount: number = 0;
+        let totalSize: number = 0;
+
+        zip.getEntries().forEach(entry => {
+            // Protect against zip bomb
+            fileCount++;
+            if (fileCount > Utils.MAX_FILES_EXTRACTED_ZIP) {
+                throw new ZipExtractError(zipPath, 'Reached max files allowed');
+            }
+
+            let entrySize: number = entry.getData().length;
+            totalSize += entrySize;
+            if (totalSize > Utils.MAX_SIZE_EXTRACTED_ZIP) {
+                throw new ZipExtractError(zipPath, 'Reached max size allowed');
+            }
+
+            let compressionRatio: number = entrySize / entry.header.compressedSize;
+            if (compressionRatio > Utils.COMPRESSION_THRESHOLD_RATIO) {
+                throw new ZipExtractError(zipPath, 'Reached max compression ratio allowed');
+            }
+
+            if (entry.isDirectory) {
+                return;
+            }
+
+            if (!zip.extractEntryTo(entry, targetDir, true, overwrite, keepOriginalPermission)) {
+                throw new ZipExtractError(zipPath, "can't extract entry " + entry.entryName);
+            }
+        });
+    }
+
+    public static extractZipEntry(zipPath: string, name: string): any {
+        if (!fs.existsSync(zipPath)) {
+            return '';
+        }
+        let zip: AdmZip = new AdmZip(zipPath);
+
+        let entry: IZipEntry | null = zip.getEntry(name);
+        if (!entry) {
+            throw new ZipExtractError(zipPath, 'Could not find expected content ' + name);
+        }
+        return entry.getData().toString('utf8');
+    }
+
     public static addWinSuffixIFNeeded(str: string): string {
         return str + (os.platform() === 'win32' ? '.exe' : '');
+    }
+}
+
+export class ZipExtractError extends Error {
+    constructor(public readonly zipPath: string, reason: string) {
+        super('Zip extraction error: ' + reason + ' in zip ' + zipPath);
     }
 }
